soc-simulator/ml_engine/requirements.txt
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
redis==5.0.1
httpx==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
torch==2.1.1
anthropic==0.7.8
joblib==1.3.2
pytest==8.3.4
